refactor(tree): shift tree manager to utils
