+++ packages/elements/src/tree/helpers/types.ts
@@ -1,29 +1,7 @@
-import type { CollectionComposer, DataItem } from '@refinitiv-ui/utils/collection.js';
+import type { CollectionComposer, TreeDataItem } from '@refinitiv-ui/utils/collection.js';
 
 export type TreeData<T extends TreeDataItem = TreeDataItem> = T[] | CollectionComposer<T> | null;
-
-export interface TreeDataItem extends DataItem {
-  /**
-   * Icon to show, when rendering the item.
-   */
-  icon?: string;
-
-  /**
-   * Label to show, when rendering the item.
-   */
-  label?: string;
-
-  /**
-   * Expanded state of child items.
-   * If `true`, child items will be visible
-   */
-  expanded?: boolean;
-
-  /**
-   * Used for nested TreeDataItem.
-   */
-  items?: TreeDataItem[];
-}
+export type { TreeDataItem } from '@refinitiv-ui/utils/collection.js';
 
 /**
  * Predicate callback
